chore: Clean up merge mistakes after the series of refactoring merges
