+++ common/src/main/java/com/wynntils/models/abilities/ShamanTotemModel.java
@@ -15,8 +15,8 @@ import com.wynntils.mc.objects.Location;
 import com.wynntils.mc.utils.McUtils;
 import com.wynntils.models.abilities.event.TotemEvent;
 import com.wynntils.models.character.event.CharacterUpdateEvent;
+import com.wynntils.models.spells.event.TrySpellCastEvent;
 import com.wynntils.models.spells.type.SpellType;
-import com.wynntils.wynn.event.TrySpellCastEvent;
 import com.wynntils.wynn.utils.WynnUtils;
 import java.util.ArrayList;
 import java.util.List;
