+++ common/src/main/java/com/wynntils/models/spells/SpellModel.java
@@ -7,12 +7,12 @@ package com.wynntils.models.spells;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Model;
 import com.wynntils.mc.event.SubtitleSetTextEvent;
+import com.wynntils.models.spells.event.SpellEvent;
+import com.wynntils.models.spells.event.SpellProgressEvent;
 import com.wynntils.models.spells.event.SpellSegmentUpdateEvent;
+import com.wynntils.models.spells.event.TrySpellCastEvent;
+import com.wynntils.models.spells.type.SpellDirection;
 import com.wynntils.models.spells.type.SpellType;
-import com.wynntils.wynn.event.SpellEvent;
-import com.wynntils.wynn.event.SpellProgressEvent;
-import com.wynntils.wynn.event.TrySpellCastEvent;
-import com.wynntils.wynn.objects.SpellDirection;
 import java.util.Arrays;
 import java.util.regex.MatchResult;
 import java.util.regex.Matcher;
