+++ common/src/main/java/com/wynntils/models/spells/event/TrySpellCastEvent.java
@@ -2,10 +2,10 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.event;
+package com.wynntils.models.spells.event;
 
+import com.wynntils.models.spells.type.SpellDirection;
 import com.wynntils.models.spells.type.SpellType;
-import com.wynntils.wynn.objects.SpellDirection;
 
 /**
  * Fired upon successful, completed spell cast.
