+++ common/src/main/java/com/wynntils/models/spells/type/SpellDirection.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.wynn.objects;
+package com.wynntils.models.spells.type;
 
 import com.wynntils.mc.utils.McUtils;
 import java.util.Arrays;
